Changes to entity type and action type. Move some shared utils. Untested, just saving my work.
