+++ api/namex/resources/auto_analyse/issues/designation_mismatch.py
@@ -2,7 +2,7 @@ from string import Template
 
 from namex.services.name_request.auto_analyse import AnalysisIssueCodes
 
-from namex.utils.service_utils import get_entity_type_description
+from namex.utils.entity_type import get_entity_type_description
 
 # Import DTOs
 from .abstract import AnalysisResponseIssue
