+++ api/namex/utils/entity_type.py
